rename to `TomlFile` and move into `kraken.common.toml`
