+++ kraken-build/src/kraken/std/python/buildsystem/maturin.py
@@ -11,9 +11,10 @@ from dataclasses import dataclass
 from pathlib import Path
 
 from kraken.common.path import is_relative_to
+from kraken.common.toml import TomlFile
 
 from ...cargo.manifest import CargoMetadata
-from ..pyproject import PyprojectHandler, TomlConfig
+from ..pyproject import PyprojectHandler
 from ..settings import PythonSettings
 from . import ManagedEnvironment
 from .pdm import PDMManagedEnvironment, PDMPythonBuildSystem
@@ -54,7 +55,7 @@ class MaturinZigTarget:
 
 class _MaturinBuilder:
     def __init__(
-        self, entry_point: str, get_pyproject_reader: Callable[[TomlConfig], PyprojectHandler], project_directory: Path
+        self, entry_point: str, get_pyproject_reader: Callable[[TomlFile], PyprojectHandler], project_directory: Path
     ) -> None:
         self._entry_point = entry_point
         self._get_pyproject_reader = get_pyproject_reader
@@ -186,13 +187,13 @@ class MaturinPoetryPythonBuildSystem(PoetryPythonBuildSystem):
 
     # PythonBuildSystem
 
-    def get_pyproject_reader(self, pyproject: TomlConfig) -> MaturinPoetryPyprojectHandler:
+    def get_pyproject_reader(self, pyproject: TomlFile) -> MaturinPoetryPyprojectHandler:
         return MaturinPoetryPyprojectHandler(pyproject)
 
     def get_managed_environment(self) -> ManagedEnvironment:
         return MaturinPoetryManagedEnvironment(self.project_directory)
 
-    def update_pyproject(self, settings: PythonSettings, pyproject: TomlConfig) -> None:
+    def update_pyproject(self, settings: PythonSettings, pyproject: TomlFile) -> None:
         super().update_pyproject(settings, pyproject)
         handler = self.get_pyproject_reader(pyproject)
         handler.synchronize_project_section_to_poetry_state()
